rename logger.js to log.js
