+++ server.js
@@ -10,7 +10,7 @@
 
 var PaymentServer = require('./lib/PaymentServer.js');
 
-var logger = require('./lib/logger.js');
+var logger = require('./lib/log.js');
 
 var payServer = new PaymentServer('.', process.env.port || 3000);
 var server = payServer.create_listener();
